Code cleanup and change port
