+++ gui/index.html
@@ -117,16 +117,16 @@
 	shutdownrpi = function () {
 		var answer = window.confirm("Shutdown ExoMy?");
 		if (answer) {
-			var surl= 'http://' + host_url + ':8099/shutdown'
+			var surl= 'http://' + host_url + ':8001/shutdown'
 
-			console.log("shutdown");
+			console.log("Shuting down started...");
 			const Http = new XMLHttpRequest();
 			Http.open("GET", surl);
 			Http.send();
 
 			Http.onreadystatechange = (e) => {
 				console.log(Http.responseText);
-				console.log("shutdown OK!");
+				console.log("Shutdown OK!");
 			}
 			alert("ExoMy shuting down safely...")
 		}
